% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_repo.R
\name{add_repo}
\alias{add_repo}
\title{Add the drat repository for the Biometry Hub packages to .Rprofile to enable easy update and installation.}
\usage{
add_repo(CRAN = "https://cloud.r-project.org/", ...)
}
\arguments{
\item{CRAN}{Default CRAN repository to add. Full list of CRAN repositories can be found at \url{https://cran.r-project.org/mirrors.html}.}

\item{...}{Other repositories to add.}
}
\description{
Add the drat repository for the Biometry Hub packages to .Rprofile to enable easy update and installation.
}
\keyword{internal}
