% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiometryTraining-deprecated.R
\name{BiometryTraining-deprecated}
\alias{BiometryTraining-deprecated}
\title{Deprecated functions in package \pkg{BiometryTraining}.}
\value{
No return value.
}
\description{
The functions listed below are deprecated and will be removed in
the future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.
}
\keyword{internal}
