% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiometryTraining-package.R
\docType{package}
\name{BiometryTraining}
\alias{BiometryTraining}
\alias{_PACKAGE}
\alias{BiometryTraining-package}
\title{BiometryTraining: Functions to Assist Design and Analysis of Agronomic Experiments}
\description{
Provides functions to aid in the design and analysis of agronomic and agricultural experiments through easy access to documentation and helper functions, especially for users who are learning these concepts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometryhub.github.io/BiometryTraining/}
  \item Report bugs at \url{https://github.com/biometryhub/BiometryTraining/issues}
}

}
\author{
\strong{Maintainer}: Sam Rogers \email{biometrytraining@adelaide.edu.au}

Authors:
\itemize{
  \item Sharon Nielsen \email{biometrytraining@adelaide.edu.au}
  \item Annie Conway \email{biometrytraining@adelaide.edu.au}
}

Other contributors:
\itemize{
  \item University of Adelaide [copyright holder, funder]
  \item GRDC [copyright holder, funder]
}

}
\keyword{internal}
