% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{install_asreml}
\alias{install_asreml}
\title{Install ASreml-R package}
\usage{
install_asreml(library = .libPaths()[1], quiet = FALSE, force = FALSE)
}
\arguments{
\item{library}{Library location to install ASreml-R. Uses first option in \code{.libPaths()} by default.}

\item{quiet}{Should package be installed quietly? Default is \code{TRUE}.}

\item{force}{Force ASreml-R to install, ignoring if it is already installed.}
}
\value{
Silently returns \code{TRUE} if \code{asreml} installed successfully or already present, \code{FALSE} otherwise. Also prints a confirmation message on success.
}
\description{
A helper function for installing the ASreml-R package, intended to reduce the difficulty of finding the correct version for your operating system and R version.
}
\keyword{internal}
