% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logltest.R
\name{logl.test}
\alias{logl.test}
\title{Log-likelihood test for comparing terms in ASReml-R models}
\usage{
logl.test(
  model.obj,
  rand.terms = NULL,
  resid.terms = NULL,
  decimals = 3,
  numeric = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{model.obj}{An ASReml-R model object}

\item{rand.terms}{Random terms from the model. Default is NULL.}

\item{resid.terms}{Residual terms from the model. Default is NULL.}

\item{decimals}{Controls rounding of decimal places in output. Default is 3 decimal places.}

\item{numeric}{Return p-values as numeric? Default is that they are characters, where very small values shown as less than a small number. See \code{details} for more.}

\item{quiet}{Logical (default: \code{FALSE}). Hide warnings and messages?}
}
\value{
A dataframe containing the results of the test.
}
\description{
Log-likelihood test for comparing terms in ASReml-R models
}
\details{
Typically p-values cannot be 0, and are usually just below some threshold of accuracy in calculation of probability.
}
\examples{
\dontrun{
library(agridat)
dat <- yates.oats
dat$row <- factor(dat$row)
dat$col <- factor(dat$col)
dat$nitro <- factor(dat$nitro)
dat$mplot <- factor(as.numeric(dat$gen))
dat <- dat[order(dat$row, dat$col), ]
library(asreml)
dat.asr <- asreml(yield ~ gen + nitro + gen:nitro,
  random = ~ block + block:mplot,
  residual = ~ ar1(row):id(col), data = dat
)
oats.logl <- logl.test(
  model.obj = dat.asr, rand.terms = c("block", "block:mplot"),
  resid.terms = c("ar1(row)")
)
oats.logl
}

}
