% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mct.R
\name{mct.out}
\alias{mct.out}
\title{Multiple Comparison Tests}
\usage{
mct.out(model.obj, pred.obj, sig = 0.95, pred, typeR, trans = NA, offset = NA)
}
\arguments{
\item{model.obj}{An ASReml-R model object.}

\item{pred.obj}{An ASReml-R prediction object with \code{sed=TRUE}.}

\item{sig}{The confidence level, numeric between 0 and 1. Default is 0.95.}

\item{pred}{Name of predictor variable as string.}

\item{typeR}{Type of test as string.}

\item{trans}{Transformation that was applied to the response variable before modelling, e.g. \code{log} or \code{sqrt}. Default is \code{NA}.}

\item{offset}{Numeric offset applied to response variable prior to transformation. Default is \code{NA}.}
}
\value{
A list containing a data frame "pred.tab" consisting of predicted means, standard errors, confidence interval upper and lower bounds, and significant group allocations.
}
\description{
A function for comparing and ranking predicted means with Tukey's Honest Significant Difference (HSD) Test.
}
\examples{
\dontrun{library(asreml)
data(oats)

#Fit ASreml Model
model.asr <- asreml(yield ~ Nitrogen*Variety,
                    random =~Blocks + Blocks:Wplots,
                    residual =~units,
                    data=oats)

wald(model.asr) #Nitrogen main effect significant

#Calculate predicted means
pred.asr <- predict(model.asr, classify="Nitrogen",sed = TRUE)

#Determine ranking and groups according to Tukey's Test
tuk.rank <- mct.out(model.obj = model.asr, pred.obj = pred.asr, sig = 0.95,
                    pred = "Nitrogen", typeR = "tukey")

tuk.rank}

}
