% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persist.R
\name{persist}
\alias{persist}
\title{Persist the Biometry Hub drat repo across sessions}
\usage{
persist(..., cran = "https://cloud.r-project.org/", profile_location = NULL)
}
\arguments{
\item{...}{Other repositories to add. These can be listed using \code{name = URL} syntax, including drat repositories.}

\item{cran}{Default CRAN repository to add. A complete list of CRAN repositories can be found at \url{https://cran.r-project.org/mirrors.html}.}

\item{profile_location}{A path to a .Rprofile file specified by the user.}
}
\description{
Add the drat repository for the Biometry Hub packages to .Rprofile to enable easy update and installation even after restating R.
}
\details{
The standard locations of .Rprofile files are searched (see \code{\link[base:Startup]{base::Startup()}}) (mostly).
}
\keyword{internal}
