% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.des.R
\name{plot.des}
\alias{plot.des}
\title{Produces graph of design layout}
\usage{
\method{plot}{des}(design.obj, nrows, ncols, brows, bcols)
}
\arguments{
\item{design.obj}{An \code{agricolae} design object.}

\item{nrows}{The number of rows in the design.}

\item{ncols}{The number of columns in the design.}

\item{brows}{For RCBD only. The number of rows in a block.}

\item{bcols}{For RCBD only. The number of columns in a block.}
}
\value{
Returns ggplot object of design layout.
}
\description{
Produces graph of design layout
}
\examples{
#This is not used by people, it forms part of des.info


}
