% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mct.R
\name{print.mct}
\alias{print.mct}
\title{Print method for mct.out objects}
\usage{
\method{print}{mct}(object, ...)
}
\arguments{
\item{object}{An mct object to print to the console.}

\item{...}{Other arguments to be passed through.}
}
\value{
The original object invisibly.
}
\description{
Print method for mct.out objects
}
\examples{
dat.aov <- aov(Petal.Width ~ Species, data = iris)
output <- mct.out(dat.aov, classify = "Species")
autoplot(output, label_height = 0.5)
}
\seealso{
\code{\link[=mct.out]{mct.out()}}
}
