% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplot.R
\name{resplot}
\alias{resplot}
\alias{resplt}
\title{Residual plots of aov or asreml models.}
\usage{
resplot(mod.obj, shapiro = TRUE, label.size = 10, axes.size = 10)

resplt(mod.obj, shapiro = TRUE, label.size = 10, axes.size = 10)
}
\arguments{
\item{mod.obj}{An \code{aov} or \code{asreml} model object.}

\item{shapiro}{(Logical) Display the Shapiro-Wilks test of normality on the plot?}

\item{label.size}{A numeric value for the size of the label (A,B,C) font point size.}

\item{axes.size}{A numeric value for the size of the axes label font size in points.}
}
\value{
A list containing ggplot2 objects which are diagnostic plots.
}
\description{
Produces plots of residuals for assumption checking of an ANOVA or asremlR model.
}
\examples{
dat.aov <- aov(Petal.Length ~ Petal.Width, data = iris)
resplot(dat.aov)
resplt(dat.aov)
}
