% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplt.R
\name{resplt}
\alias{resplt}
\title{Produces plots of residuals for assumption checking of an ANOVA model.}
\usage{
resplt(mod.obj)
}
\arguments{
\item{mod.obj}{An \code{aov} model.}
}
\value{
A list containing ggplot2 objects which are ANOVA diagnostic plots.
}
\description{
Produces plots of residuals for assumption checking of an ANOVA model.
}
\examples{
dat.aov <- aov(Petal.Length ~ Petal.Width, data = iris)
resplt(dat.aov)

}
