% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuk.out.R
\name{tuk.out-deprecated}
\alias{tuk.out-deprecated}
\title{Tukey's Honest Significant Difference test (Deprecated)}
\usage{
tuk.out(pred.obj, model.obj, pred, sig = 0.95, trans = FALSE, offset = 0)
}
\arguments{
\item{pred.obj}{An ASReml-R prediction object.}

\item{model.obj}{An ASReml-R model object.}

\item{pred}{A character string of the main effect or interaction terms predicted from the model.}

\item{sig}{The significance level to test at.}

\item{trans}{Choice of \code{log} or \code{sqrt} to back-transform predicted values and standard errors in the output.}

\item{offset}{A numeric value used in the transformation.}
}
\value{
A data frame with the predicted values and results of Tukey's test.
}
\description{
This function has been replaced with \code{\link[=mct.out]{mct.out()}} and will be removed in a future version of this package. Performs Tukey's Honest Significant Difference test and optionally back-transforms natural log and squareroot transformed predicted values.
}
\examples{
\dontrun{
dat.asr <- asreml(Yield ~ Variety,
  random = ~Block,
  residual = ~ id(Plot), data = dat
)

dat.pred <- predict(dat.asr,
  classify = "Variety",
  sed = TRUE
)

pred.out <- tuk.out(
  model.obj = dat.asr, pred.obj = dat.pred,
  pred = "Variety", sig = 0.95
)

pred.out
}

}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.47.9023}
}
\seealso{
\code{\link{BiometryTraining-deprecated}}
}
\keyword{internal}
