% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.rank.R
\name{tukey.rank}
\alias{tukey.rank}
\title{Compares means pairwise using Tukey's Honest Significant Difference test}
\usage{
tukey.rank(Mean, SED, Names = NULL, crit.val)
}
\arguments{
\item{Mean}{The predicted mean from an \code{ASreml-R} \code{predict} function call.}

\item{SED}{The Standard Error of Difference from an \code{ASreml-R} \code{predict} function call.}

\item{Names}{The treatment names.}

\item{crit.val}{The confidence level required.}
}
\value{
A data frame with the predicted means, standard errors and Tukey's rankings.
}
\description{
Compares means pairwise using Tukey's Honest Significant Difference test
}
\examples{
#This is not used by people, it forms part of tuk.out

}
\author{
Remy van de Ven
}
