% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{update_asreml}
\alias{update_asreml}
\title{Update asreml package}
\usage{
update_asreml(...)
}
\arguments{
\item{...}{other arguments passed to \code{\link[=install_asreml]{install_asreml()}}}
}
\description{
Update asreml package
}
