% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram.r
\name{vario}
\alias{vario}
\title{Variogram plots for spatial models.}
\usage{
vario(mod.obj)
}
\arguments{
\item{mod.obj}{An \code{asreml} model object.}
}
\value{
A list containing ggplot2 objects.
}
\description{
Produces variogram plots for checking spatial trends.
}
\examples{
\dontrun{
library(asreml)
oats <- asreml::oats
oats <- oats[order(oats$Row, oats$Column),]
model.asr <- asreml(yield ~ Nitrogen + Variety + Nitrogen:Variety,
                    random = ~ Blocks + Blocks:Wplots,
                    residual = ~ ar1(Row):ar1(Column),
                    data = oats)
vario(model.asr)
}
}
