% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot.mct}
\alias{autoplot.design}
\title{Automatic plots for objects}
\usage{
\method{autoplot}{mct}(object, label_height = 0.1, rotation = 0, ...)

\method{autoplot}{design}(design)
}
\arguments{
\item{object}{A data frame of output from the \code{\link[=mct.out]{mct.out()}} function with class "mct".}

\item{label_height}{Height of the text labels above the upper error bar on the plot. Default is 0.1 (10\%) of the difference between upper and lower error bars above the top error bar.}

\item{rotation}{Rotate the text output as Treatments within the plot. Allows for easier reading of long treatment labels. Number between 0 and 360 (inclusive) - default 0}

\item{...}{Other arguments to be passed through.}
}
\value{
A \code{ggplot2} object.
}
\description{
Automatic plots for objects
}
\examples{
dat.aov <- aov(Petal.Width ~ Species, data = iris)
output <- mct.out(dat.aov, classify = "Species")
autoplot(output, label_height = 0.25)
}
\seealso{
\code{\link[=mct.out]{mct.out()}} and \code{\link[=design]{design()}}
}
