% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{autoplot.mct}
\alias{autoplot.design}
\title{Objects exported from other packages}
\usage{
\method{autoplot}{mct}(object, rotation = 0, size = 4, label_height = 0.1, ...)

\method{autoplot}{design}(object, rotation = 0, size = 4, margin = FALSE, colour_blind = F, ...)
}
\arguments{
\item{object}{An object to create a plot for. Currently objects rom the \code{\link[=mct.out]{mct.out()}} or \code{\link[=design]{design()}} functions with class "mct" or "design" respectively are supported.}

\item{rotation}{Rotate the text output as Treatments within the plot. Allows for easier reading of long treatment labels. Number between 0 and 360 (inclusive) - default 0}

\item{size}{Increase or decrease the text size within the plot for treatment labels. Numeric with default value of 4.}

\item{label_height}{Height of the text labels above the upper error bar on the plot. Default is 0.1 (10\%) of the difference between upper and lower error bars above the top error bar.}

\item{...}{Other arguments to be passed through.}

\item{margin}{Logical (default \code{FALSE}). A value of \code{FALSE} will expand the plot to the edges of the plotting area i.e. remove white space between plot and axes.}
}
\value{
A \code{ggplot2} object.
}
\examples{
dat.aov <- aov(Petal.Width ~ Species, data = iris)
output <- mct.out(dat.aov, classify = "Species")
autoplot(output, label_height = 0.5)
}
\seealso{
\code{\link[=mct.out]{mct.out()}} and \code{\link[=design]{design()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

