% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiometryTraining-package.R
\docType{package}
\name{BiometryTraining}
\alias{BiometryTraining}
\alias{BiometryTraining-package}
\title{BiometryTraining: Functions to Assist Teaching Design and Analysis of
    Agronomic Experiments}
\description{
BiometryTraining gives users a simple way to install and 
    access documentation about the functions used and distributed as 
    part of the SAGI-STH Biometrics Education Initiative.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometryhub.github.io/BiometryTraining/}
  \item Report bugs at \url{https://github.com/biometryhub/BiometryTraining/issues}
}

}
\author{
\strong{Maintainer}: Sam Rogers \email{biometrytraining@adelaide.edu.au}

Authors:
\itemize{
  \item Sharon Nielsen \email{biometrytraining@adelaide.edu.au}
  \item Annie Conway \email{biometrytraining@adelaide.edu.au}
}

}
