% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.info.R
\name{des.info}
\alias{des.info}
\title{Produces graph of design layout, skeletal ANOVA table and data frame with complete design}
\usage{
des.info(design.obj, nrows, ncols, brows = NA, bcols = NA,
  rotation = 0, size = 4, margin = TRUE, quiet = FALSE)
}
\arguments{
\item{design.obj}{An \code{agricolae} design object.}

\item{nrows}{The number of rows in the design.}

\item{ncols}{The number of columns in the design.}

\item{brows}{For RCBD only. The number of rows in a block.}

\item{bcols}{For RCBD only. The number of columns in a block.}

\item{rotation}{Rotate the text output as Treatments within the plot. Allows for easier reading of long treatment labels.}

\item{size}{Increase or decrease the text size within the plot for treatment labels. Numeric with default value of 4.}

\item{margin}{Logical (default FALSE). Expand the plot to the edges of the plotting area i.e. remove whitespace between plot and axes.}

\item{quiet}{Logical (default FALSE). Return the objects without printing output.}
}
\value{
A list containing a data frame with the complete design and a ggplot object with plot layout.
}
\description{
Produces graph of design layout, skeletal ANOVA table and data frame with complete design
}
\examples{
library(agricolae)
trt <- c(1,5,10,20)
rep <- 5
outdesign <- design.crd(trt = trt, r=rep)
des.out <- des.info(design.obj = outdesign, nrows = 4,
                    ncols = 5, brows = NA, bcols = NA)

trt <- LETTERS[1:11]
rep <- 4
outdesign <- design.rcbd(trt = trt, r = rep)
des.out <- des.info(design.obj = outdesign, nrows = 11,
                    ncols = 4, brows = 11, bcols = 1)

}
