% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{install_asreml}
\alias{install_asreml}
\title{Install Asreml package}
\usage{
install_asreml(library = .libPaths()[1], quiet = TRUE)
}
\arguments{
\item{library}{Library location to install ASReml-R. Uses default library location by default.}

\item{quiet}{Should package be installed quietly? Default is \code{TRUE}.}
}
\value{
\code{TRUE} if asreml installed successfully or already present, \code{FALSE} otherwise. Also prints a confirmation message on success.
}
\description{
Install Asreml package
}
\keyword{internal}
