% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_asreml.R
\name{install_asreml}
\alias{install_asreml}
\title{Install ASreml-R package}
\usage{
install_asreml(
  library = .libPaths()[1],
  quiet = FALSE,
  force = FALSE,
  keep_file = FALSE
)
}
\arguments{
\item{library}{Library location to install ASreml-R. Uses first option in \code{.libPaths()} by default.}

\item{quiet}{Logical (default \code{FALSE}). Should package be installed quietly?}

\item{force}{Logical (default \code{FALSE}). Force ASreml-R to install. Useful for upgrading if it is already installed.}

\item{keep_file}{Should the downloaded asreml package file be kept? Default is \code{FALSE}. \code{TRUE} downloads to current directory. Can also provide a file path to save to another directory.}
}
\value{
Silently returns \code{TRUE} if \code{asreml} installed successfully or already present, \code{FALSE} otherwise. Optionally prints a confirmation message on success.
}
\description{
A helper function for installing the ASreml-R package, intended to reduce the difficulty of finding the correct version for your operating system and R version.
}
