% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logltest.R
\name{logl.test}
\alias{logl.test}
\title{Log-likelihood test for comparing terms in ASreml-R models}
\usage{
logl.test(model.obj, rand.terms, resid.terms)
}
\arguments{
\item{model.obj}{An ASreml-R model object}

\item{rand.terms}{Random terms from the model}

\item{resid.terms}{Residual terms from the model}
}
\value{
A dataframe containing the results of the test.
}
\description{
Log-likelihood test for comparing terms in ASreml-R models
}
