% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logltest.R
\name{logl.test}
\alias{logl.test}
\title{Log-likelihood test for comparing terms in ASreml-R models}
\usage{
logl.test(model.obj, rand.terms, resid.terms)
}
\arguments{
\item{model.obj}{An ASreml-R model object}

\item{rand.terms}{Random terms from the model}

\item{resid.terms}{Residual terms from the model}
}
\value{
A dataframe containing the results of the test.
}
\description{
Log-likelihood test for comparing terms in ASreml-R models
}
\details{
Uses the \code{\link[asremlPlus:REMLRT.asreml]{asremlPlus::REMLRT.asreml()}} function to calculate the Likelihood ratio test for Asreml-R objects.
}
\examples{
\dontrun{
library(agridat)
dat <- yates.oats
dat$row <- factor(dat$row)
dat$col <- factor(dat$col)
dat$nitro <- factor(dat$nitro)
dat$mplot <- factor(as.numeric(dat$gen))
dat <- dat[order(dat$row, dat$col), ]
library(asreml)
dat.asr <- asreml(yield ~ gen + nitro + gen:nitro,
  random = ~ block + block:mplot,
  residual = ~ ar1(row):id(col), data = dat
)
oats.logl <- logl.test(
  model.obj = dat.asr, rand.terms = c("block", "block:mplot"),
  resid.terms = c("ar1(row)")
)
oats.logl
}

}
