% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.des.R
\name{plot.des}
\alias{plot.des}
\title{Produces graph of design layout}
\usage{
\method{plot}{des}(
  design.obj,
  nrows,
  ncols,
  brows,
  bcols,
  byrow,
  rotation,
  size,
  margin,
  return.seed,
  fac.sep
)
}
\arguments{
\item{design.obj}{An \code{agricolae} design object.}

\item{nrows}{The number of rows in the design.}

\item{ncols}{The number of columns in the design.}

\item{brows}{For RCBD only. The number of rows in a block.}

\item{bcols}{For RCBD only. The number of columns in a block.}

\item{byrow}{For split-plot only. Logical. Provides a way to arrange plots within whole-plots when there are multiple possible arrangements.}

\item{rotation}{Rotate the text output as Treatments within the plot. Allows for easier reading of long treatment labels. Number between 0 and 360 (inclusive) - default 0}

\item{size}{Increase or decrease the text size within the plot for treatment labels. Numeric with default value of 4.}

\item{margin}{Logical (default FALSE). Expand the plot to the edges of the plotting area i.e. remove whitespace between plot and axes.}

\item{return.seed}{Logical (default TRUE). Output the seed used in the design?}

\item{fac.sep}{Character to separate factorial treatments.}

\item{quiet}{Logical (default FALSE). Return the objects without printing output.}
}
\value{
Returns dataframe of design and ggplot object of design layout.
}
\description{
Produces graph of design layout
}
\keyword{internal}
