% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mct.R, R/print.R
\name{print.mct}
\alias{print.mct}
\title{Print method for mct.out objects}
\usage{
\method{print}{mct}(object, ...)

\method{print}{mct}(object, ...)
}
\arguments{
\item{object}{An object to print to the console.for. Currently objects rom the \code{\link[=mct.out]{mct.out()}} or \code{\link[=design]{design()}} functions with class "mct" or "design" respectively are supported.}

\item{...}{Other arguments to be passed through.}
}
\value{
The original object invisibly.

The original object invisibly.
}
\description{
Print method for mct.out objects

Print method for mct.out objects
}
\examples{
dat.aov <- aov(Petal.Width ~ Species, data = iris)
output <- mct.out(dat.aov, classify = "Species")
autoplot(output, label_height = 0.5)
dat.aov <- aov(Petal.Width ~ Species, data = iris)
output <- mct.out(dat.aov, classify = "Species")
autoplot(output, label_height = 0.5)
}
\seealso{
\code{\link[=mct.out]{mct.out()}}

\code{\link[=mct.out]{mct.out()}} and \code{\link[=design]{design()}}
}
