% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resplt.R
\name{resplot}
\alias{resplot}
\alias{resplt}
\title{Residual plots of aov or asreml models.}
\usage{
resplot(mod.obj, shapiro = TRUE)
}
\arguments{
\item{mod.obj}{An \code{aov} or \code{asreml} model object.}

\item{shapiro}{(Logical) Display the Shapiro-Wilks test of normality on the plot?}
}
\value{
A list containing ggplot2 objects which are diagnostic plots.
}
\description{
Produces plots of residuals for assumption checking of an ANOVA or asremlR model.
}
\examples{
dat.aov <- aov(Petal.Length ~ Petal.Width, data = iris)
resplot(dat.aov)
resplt(dat.aov)
}
