% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram.r
\name{variogram}
\alias{variogram}
\title{Variogram plots for spatial models.}
\usage{
variogram(model.obj, row = NA, column = NA)
}
\arguments{
\item{model.obj}{An \code{asreml} model object.}

\item{row}{A row variable.}

\item{column}{A column variable.}
}
\value{
A list containing ggplot2 objects.
}
\description{
Produces variogram plots for checking spatial trends.
}
\examples{
\dontrun{
library(asreml)
oats <- asreml::oats
oats <- oats[order(oats$Row, oats$Column),]
model.asr <- asreml(yield ~ Nitrogen + Variety + Nitrogen:Variety,
                    random = ~ Blocks + Blocks:Wplots,
                    residual = ~ ar1(Row):ar1(Column),
                    data = oats)
variogram(model.asr)
}
}
\references{
S. P. Kaluzny, S. C. Vega, T. P. Cardoso, A. A. Shelly, \emph{S+SpatialStats: User’s Manual for Windows® and UNIX®} (Springer New York, 2013; https://books.google.com.au/books?id=iADkBwAAQBAJ).
}
