% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram.r
\name{variogram}
\alias{variogram}
\title{Calculate the variogram for a model}
\usage{
variogram(model.obj)
}
\arguments{
\item{model.obj}{An asreml model}
}
\value{
A data frame with the variogram for a model. The data frame contains the spatial coordinaties (typically row and column), the $gamma$ for that position and the number of points with the separation.
}
\description{
Calculate the variogram for a model
}
\examples{
\dontrun{
library(asreml)
oats <- asreml::oats
oats <- oats[order(oats$Row, oats$Column),]
model.asr <- asreml(yield ~ Nitrogen + Variety + Nitrogen:Variety,
                    random = ~ Blocks + Blocks:Wplots,
                    residual = ~ ar1(Row):ar1(Column),
                    data = oats)
variogram(model.asr)
}

}
\references{

}
\keyword{internal}
